requirements.txt
python-dotenv>=1.0.0
numpy>=1.26.0
typer>=0.12.0
rich>=13.7.0
python-dateutil>=2.9.0
